Move JoinType enum from spilling LookupJoinOperatorFactory
It's shared between spilling and non-spilling Join operators. Moved to
shared place.
