+++ core/trino-main/src/main/java/io/trino/operator/join/DefaultPageJoiner.java
@@ -26,7 +26,6 @@ import io.trino.operator.WorkProcessor;
 import io.trino.operator.exchange.LocalPartitionGenerator;
 import io.trino.operator.join.spilling.JoinProbe;
 import io.trino.operator.join.spilling.JoinProbe.JoinProbeFactory;
-import io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType;
 import io.trino.operator.join.spilling.LookupJoinPageBuilder;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
@@ -57,8 +56,8 @@ import static io.trino.operator.WorkProcessor.TransformationState.finished;
 import static io.trino.operator.WorkProcessor.TransformationState.needsMoreData;
 import static io.trino.operator.WorkProcessor.TransformationState.ofResult;
 import static io.trino.operator.WorkProcessor.TransformationState.yielded;
-import static io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType.FULL_OUTER;
-import static io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType.PROBE_OUTER;
+import static io.trino.operator.join.JoinType.FULL_OUTER;
+import static io.trino.operator.join.JoinType.PROBE_OUTER;
 import static io.trino.operator.join.spilling.PartitionedLookupSourceFactory.NO_SPILL_EPOCH;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
