+++ core/trino-main/src/main/java/io/trino/operator/join/JoinStatisticsCounter.java
@@ -14,7 +14,6 @@
 package io.trino.operator.join;
 
 import io.trino.operator.OperatorInfo;
-import io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType;
 
 import java.util.OptionalLong;
 import java.util.function.Supplier;
