+++ core/trino-main/src/main/java/io/trino/operator/join/JoinType.java
@@ -1,0 +1,22 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.operator.join;
+
+public enum JoinType
+{
+    INNER,
+    PROBE_OUTER, // the Probe is the outer side of the join
+    LOOKUP_OUTER, // The LookupSource is the outer side of the join
+    FULL_OUTER,
+}
