+++ core/trino-main/src/main/java/io/trino/operator/join/spilling/LookupJoinOperator.java
@@ -22,11 +22,11 @@ import io.trino.operator.WorkProcessor;
 import io.trino.operator.WorkProcessorOperator;
 import io.trino.operator.join.DefaultPageJoiner;
 import io.trino.operator.join.JoinStatisticsCounter;
+import io.trino.operator.join.JoinType;
 import io.trino.operator.join.LookupSourceFactory;
 import io.trino.operator.join.LookupSourceProvider;
 import io.trino.operator.join.PageJoiner.PageJoinerFactory;
 import io.trino.operator.join.spilling.JoinProbe.JoinProbeFactory;
-import io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType;
 import io.trino.spi.Page;
 import io.trino.spi.metrics.Metrics;
 import io.trino.spi.type.Type;
