+++ core/trino-main/src/main/java/io/trino/operator/join/unspilled/LookupJoinOperator.java
@@ -20,8 +20,8 @@ import io.trino.operator.ProcessorContext;
 import io.trino.operator.WorkProcessor;
 import io.trino.operator.WorkProcessorOperator;
 import io.trino.operator.join.JoinStatisticsCounter;
+import io.trino.operator.join.JoinType;
 import io.trino.operator.join.LookupSource;
-import io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType;
 import io.trino.operator.join.unspilled.JoinProbe.JoinProbeFactory;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
