+++ core/trino-main/src/main/java/io/trino/operator/join/unspilled/PageJoiner.java
@@ -19,8 +19,8 @@ import io.trino.operator.DriverYieldSignal;
 import io.trino.operator.ProcessorContext;
 import io.trino.operator.WorkProcessor;
 import io.trino.operator.join.JoinStatisticsCounter;
+import io.trino.operator.join.JoinType;
 import io.trino.operator.join.LookupSource;
-import io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType;
 import io.trino.operator.join.unspilled.JoinProbe.JoinProbeFactory;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
@@ -38,8 +38,8 @@ import static io.trino.operator.WorkProcessor.TransformationState.finished;
 import static io.trino.operator.WorkProcessor.TransformationState.needsMoreData;
 import static io.trino.operator.WorkProcessor.TransformationState.ofResult;
 import static io.trino.operator.WorkProcessor.TransformationState.yielded;
-import static io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType.FULL_OUTER;
-import static io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType.PROBE_OUTER;
+import static io.trino.operator.join.JoinType.FULL_OUTER;
+import static io.trino.operator.join.JoinType.PROBE_OUTER;
 import static java.util.Objects.requireNonNull;
 
 public class PageJoiner
