+++ core/trino-main/src/test/java/io/trino/operator/join/TestJoinOperatorInfo.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.Test;
 import java.util.OptionalLong;
 
 import static com.google.common.base.Preconditions.checkArgument;
-import static io.trino.operator.join.spilling.LookupJoinOperatorFactory.JoinType.INNER;
+import static io.trino.operator.join.JoinType.INNER;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class TestJoinOperatorInfo
